Fix UBSan error: match QuicConnRecvResumptionTicket signature to callback type
The TicketLength parameter was declared as uint16_t in the function
definition but the CXPLAT_TLS_RECEIVE_TICKET_CALLBACK typedef specifies
uint32_t. This mismatch caused an undefined behavior sanitizer error
when the function was called through the callback pointer.

+++ src/core/connection.c
@@ -2110,7 +2110,7 @@ _IRQL_requires_max_(PASSIVE_LEVEL)
 BOOLEAN
 QuicConnRecvResumptionTicket(
     _In_ QUIC_CONNECTION* Connection,
-    _In_ uint16_t TicketLength,
+    _In_ uint32_t TicketLength,
     _In_reads_(TicketLength)
         const uint8_t* Ticket
     )
